requirements.txt
Flask==2.0.2
Flask-PyMongo==2.3.0
pymongo==3.12.1
Werkzeug==2.0
pymongo[srv]
python-dotenv==0.19.2
pytest==8.2.2
mongomock==3.23.0